feat: Add option to remove Ceph repo, support Proxmox 8
This adds a task and related variable to remove the Ceph Apt repository
file at /etc/apt/sources.list.d/ceph.list. I believe this only exists on
PVE 8/Debian 12 but currently causes the add-no-subscription task to
fail as mentioned in
https://github.com/ironicbadger/ansible-role-proxmox-nag-removal/issues/21.

Note: I've made no attempt to support the no subscription version of the
Ceph repository as I expect that most people interested in this are not
using Ceph anyway.

Fixes: #21